feat(router): providing url parameter is now optional
This is applicable to all helpers (get, post, head, etc.).
If url is not provided, it is referred to as empty string ('') meaning that all routes of prefixed router registered without url will have the same exact url as the prefix states.
NOTE: this will not work if there is no url neither in the route nor in the prefix. To save yourself, only use url-free declarations in prefixed routers.

BREAKING CHANGE: router.register now requires methods as the first parameter insead of the second as it was before